dev: reworks PCI to add a PCI host bridge
This commit reworks the PCI architecture to:

  - Add a PCI host bridge
  - Add the possibility to have different upstream between multiple
    `PciDevice` (for future PCI-PCI bridge implementation)
  - Moves the responsibility to response a PCI configuration request in
    `PciDevice` instead of `PciHost`.

For that, the `PciUpstream` class is added and works like the old
`PciHost`, which now inherit from it. `PciDevice` will now need to
register to a `PciUpstream` instead of `PciHost` to get an instance of
`DeviceInterface`, this allows for the future implementation of a
PCI-to-PCI bridge. Also, bus number is now given to the device by its
upstream and not by the gem5 configuration which allow it to be changed
by the OS on boot for PCI-PCI bridge.

The new `PciHost` implementation now as a member of the new class
`PciHostBridge`. That class acts like a two-way bridge between the
system bus and the PCI bus. It exposes all PCI address to the system
side and should act as the default destination on the PCI side.

The python `PciXBar` class is also added, which for now simply inherit
from `NoncoherentXBar`.

So, the new PCI architecture is composed of a `PciHost` with internal
`PciHostBridge`, connected to a `PciXBar` on PCI side to which all
`PciDevice` are connected.

PCI configuration request are now handled by each `PciDevice` instead
of `PciHost` to mimic real PCI configuration.